Add files via upload
There was an issue with this file when you first uploaded it.  The name of the intake subsystem was linked with its possible use as a hoist by means of the "/" sign.  This was apparently taken by RobotBuilder to be an indication of a subdirectory under the subsystems.  As a result the file was largely unuseable.  Something also happened when you set up the Command Group as the Autonomous Command.